+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Views
 
 Models
 
+- `OldValueCache` property wrapper is added that stores value and caches old value
 - `Clamped` property wrapper now conforms to `DynamicProperty`
 - `stackView(alignmentHor:alignmentVer:spacing:content)` method is added to `LayoutDirectionOmni`
 - `TextLineType` and `TextLineLimitType` are now exposed to APIs below `iOS` `16.0`, `macOS` `13.0`, `tvOS` `16.0`, `watchOS` `9.0`